Combined PupilineTests 1-5, 7, 9, 10, 13-16, 26. Removed some duplicates
