--- src/test/java/school/redrover/Pipeline4Test.java
@@ -1,30 +1,0 @@
-package school.redrover;
-
-import org.openqa.selenium.By;
-import org.testng.Assert;
-import org.testng.annotations.Ignore;
-import org.testng.annotations.Test;
-import school.redrover.runner.BaseTest;
-
-public class Pipeline4Test extends BaseTest {
-
-    @Ignore
-    @Test
-    public void testCreatePipelineWithValidName() {
-        final String projectName = "PipelineProjectName";
-
-        getDriver().findElement(By.xpath("//a[@href='newJob']")).click();
-        getDriver().findElement(By.id("name")).sendKeys(projectName);
-        getDriver().findElement(By.className("org_jenkinsci_plugins_workflow_job_WorkflowJob")).click();
-        getDriver().findElement(By.id("ok-button")).click();
-        getDriver().findElement(By.name("Submit")).click();
-
-        getDriver().findElement(By.id("jenkins-name-icon")).click();
-
-        getDriver().findElement(By.xpath("//td/a[@href = 'job/" + projectName + "/']")).click();
-
-        Assert.assertEquals(getDriver().findElement(By.xpath("//div[@id='main-panel']/h1")).getText(),
-                String.format("Pipeline %s", projectName));
-    }
-}
-
