+++ src/test/java/school/redrover/PipelineTest.java
@@ -16,7 +16,10 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 public class PipelineTest extends BaseTest {
-
+    private static final String JOB_NAME = "NewPipeline";
+    private static final String JOB_ON_DASHBOARD_XPATH = "//tr[@id ='job_" + JOB_NAME + "']//a[@href = 'job/" + JOB_NAME + "/']";
+    private static final String CONFIGURE_ON_SIDE_PANEL_XPATH = "//div[@id = 'tasks']//a[@href = '/job/" + JOB_NAME + "/configure']";
+    private static final String CHECKBOX_TEXT = "Do not allow concurrent build";
     private final String PIPELINE_NAME = "Pipeline test";
 
     private void createPipeline(String pipelineName, boolean returnToDashboard) {
@@ -28,6 +31,16 @@ public class PipelineTest extends BaseTest {
         if (returnToDashboard) {
             goToDashboard();
         }
+    }
+    private void createPipeline (String pipelineName) {
+        getDriver().findElement(By.xpath("//a[@href = '/view/all/newJob']")).click();
+        getDriver().findElement(By.xpath("//input[@class = 'jenkins-input']")).sendKeys(pipelineName);
+        getDriver().findElement(By.className("org_jenkinsci_plugins_workflow_job_WorkflowJob")).click();
+
+        getDriver().findElement(By.id("ok-button")).click();
+        getDriver().findElement(By.xpath("//button[@name = 'Submit']")).click();
+
+        getDriver().findElement(By.id("jenkins-home-link")).click();
     }
 
     private void clickSaveConfiguration() {
@@ -53,20 +66,41 @@ public class PipelineTest extends BaseTest {
     }
 
     @Test
-    public void testCreateWithValidName() {
-        final String pipelineName = "PipelineName";
-
-        getDriver().findElement(By.xpath("//a[@href = '/view/all/newJob']")).click();
-        getDriver().findElement(By.className("jenkins-input")).sendKeys(pipelineName);
-        getDriver().findElement(By.className("org_jenkinsci_plugins_workflow_job_WorkflowJob")).click();
-        getDriver().findElement(By.xpath("//button[@id = 'ok-button']")).click();
-
-        getDriver().findElement(By.xpath("//button[@name = 'Submit']")).click();
+    public void testCreatePipeline() {
+        WebElement newItem = getDriver().findElement(By.xpath("//a[@href='/view/all/newJob']"));
+        newItem.click();
+        WebElement itemName = getDriver().findElement(By.id("name"));
+        itemName.sendKeys(PIPELINE_NAME);
+        WebElement pipeLine = getDriver().findElement(By.className("org_jenkinsci_plugins_workflow_job_WorkflowJob"));
+        pipeLine.click();
+        WebElement button = getDriver().findElement(By.id("ok-button"));
+        button.click();
         getDriver().findElement(By.id("jenkins-name-icon")).click();
+        getDriver().findElement(By.xpath("//td/a[@href = 'job/" + PIPELINE_NAME.replace(" ", "%20") + "/']")).click();
 
-        getDriver().findElement(By.xpath("//td/a[@href = 'job/" + pipelineName + "/']")).click();
+        Assert.assertEquals(
+                getDriver().findElement(By.cssSelector("#main-panel > h1")).getText(),
+                "Pipeline " + PIPELINE_NAME);
+    }
 
-        Assert.assertEquals(getDriver().getTitle(), pipelineName + " [Jenkins]");
+    @Test
+    public void testCreatePipelineValidName() {
+        final String validPipelineName = "NewPipeline";
+        getDriver().findElement(By.xpath("//a[@href = '/view/all/newJob']")).click();
+
+        getDriver().findElement(By.xpath("//input[@id ='name']")).sendKeys(validPipelineName);
+        getDriver().findElement(By.xpath("//span[normalize-space()='Pipeline']")).click();
+        getDriver().findElement(By.xpath("//button[@id='ok-button']")).click();
+
+        getDriver().findElement(By.xpath("//button[@name='Submit']")).click();
+        getDriver().findElement(By.xpath("//img[@id='jenkins-name-icon']")).click();
+
+        getDriver().findElement(By.xpath("//td//a[@href = 'job/" + validPipelineName + "/']")).click();
+
+        Assert.assertEquals(
+                getDriver().findElement(By.xpath("//div[@id = 'main-panel']/h1")).getText(),
+                "Pipeline " + validPipelineName);
+
     }
 
     @Test
@@ -80,6 +114,23 @@ public class PipelineTest extends BaseTest {
                 "» This field cannot be empty, please enter a valid name");
         Assert.assertTrue(
                 getDriver().findElement(By.cssSelector(".disabled")).isDisplayed());
+    }
+
+    @Test
+    public void testPipelineEmptyNameHandling() {
+
+        final String errorTextExpected = "» This field cannot be empty, please enter a valid name";
+
+        getDriver().findElement(By.xpath("//a[@href = '/view/all/newJob']")).click();
+        getDriver().findElement(By.xpath("//div[@id ='j-add-item-type-standalone-projects']//span[contains(text(), 'Pipeline')]")).click();
+
+        WebElement error = getDriver().findElement(By.id("itemname-required"));
+        String errorTextActual =  error.getText();
+        String errorTextColor = error.getCssValue("color");
+
+        Assert.assertEquals(errorTextActual, errorTextExpected);
+        Assert.assertEquals(errorTextColor, "rgba(255, 0, 0, 1)");
+        Assert.assertEquals(getDriver().findElement(By.id("ok-button")).getAttribute("disabled"), "true");
     }
 
     @Test
@@ -100,25 +151,6 @@ public class PipelineTest extends BaseTest {
 
         Assert.assertTrue(getDriver().findElement(By.cssSelector("#main-panel")).getText().contains("A job already exists with the name ‘" + pipelineName + "’"));
 
-    }
-
-    @Test
-    public void testRenamePipeline() {
-        String pipelineNameForCreate = "Hello";
-        String pipelineNameForRename = "HaveAGoodDay";
-
-        createPipeline(pipelineNameForCreate, true);
-
-        getDriver().findElement(By.xpath("//*[@id='job_" + pipelineNameForCreate + "']/td[3]/a")).click();
-        getDriver().findElement(By.xpath("//a[@href = '/job/" + pipelineNameForCreate + "/confirm-rename']")).click();
-
-        getDriver().findElement(By.cssSelector(".jenkins-input.validated")).clear();
-        getDriver().findElement(By.cssSelector(".jenkins-input.validated")).sendKeys(pipelineNameForRename);
-        getDriver().findElement(By.xpath("//button[@name='Submit']")).click();
-
-        getDriver().findElement(By.xpath("//a[@id= 'jenkins-home-link']")).click();
-
-        Assert.assertTrue(getDriver().findElement(By.xpath("//td/a[@href='job/" + pipelineNameForRename + "/']")).isDisplayed());
     }
 
     @Test
@@ -183,55 +215,6 @@ public class PipelineTest extends BaseTest {
                 "No name is specified");
     }
 
-    @Test
-    public void testCreatePipeline() {
-        WebElement newItem = getDriver().findElement(By.xpath("//a[@href='/view/all/newJob']"));
-        newItem.click();
-        WebElement itemName = getDriver().findElement(By.id("name"));
-        itemName.sendKeys(PIPELINE_NAME);
-        WebElement pipeLine = getDriver().findElement(By.className("org_jenkinsci_plugins_workflow_job_WorkflowJob"));
-        pipeLine.click();
-        WebElement button = getDriver().findElement(By.id("ok-button"));
-        button.click();
-        getDriver().findElement(By.id("jenkins-name-icon")).click();
-        getDriver().findElement(By.xpath("//td/a[@href = 'job/" + PIPELINE_NAME.replace(" ", "%20") + "/']")).click();
-
-        Assert.assertEquals(
-                getDriver().findElement(By.cssSelector("#main-panel > h1")).getText(),
-                "Pipeline " + PIPELINE_NAME);
-    }
-
-    @Test
-    public void testCreatePipelineValidName() {
-        final String validPipelineName = "NewPipeline";
-        getDriver().findElement(By.xpath("//a[@href = '/view/all/newJob']")).click();
-
-        getDriver().findElement(By.xpath("//input[@id ='name']")).sendKeys(validPipelineName);
-        getDriver().findElement(By.xpath("//span[normalize-space()='Pipeline']")).click();
-        getDriver().findElement(By.xpath("//button[@id='ok-button']")).click();
-
-        getDriver().findElement(By.xpath("//button[@name='Submit']")).click();
-        getDriver().findElement(By.xpath("//img[@id='jenkins-name-icon']")).click();
-
-        getDriver().findElement(By.xpath("//td//a[@href = 'job/" + validPipelineName + "/']")).click();
-
-        Assert.assertEquals(
-                getDriver().findElement(By.xpath("//div[@id = 'main-panel']/h1")).getText(),
-                "Pipeline " + validPipelineName);
-
-    }
-
-    @Test
-    public void testCreatePipelineWithValidName() {
-        final String PipelineName = "PipelineProjectName";
-        getDriver().findElement(By.cssSelector("a[href*='/view/all/newJob']")).click();
-        getDriver().findElement(By.cssSelector("input[class ='jenkins-input']")).sendKeys(PipelineName);
-        getDriver().findElement(By.xpath("//span[normalize-space()='Pipeline']")).click();
-        getDriver().findElement(By.cssSelector("button[type = 'submit']")).click();
-        getDriver().findElement(By.cssSelector("button[name = 'Submit']")).click();
-        getDriver().findElement(By.cssSelector("li[class = 'jenkins-breadcrumbs__list-item']")).click();
-        Assert.assertEquals(getDriver().findElement(By.cssSelector("a[href = 'job/PipelineProjectName/']")).getText(), PipelineName);
-    }
 
     @Test
     public void testCreatePipelineProject() {
@@ -394,5 +377,251 @@ public class PipelineTest extends BaseTest {
                 .toList();
 
         Assert.assertEquals(buildParameters, parameterChoices);
+    }
+
+    private void createAPipeline(String jobName) {
+        getDriver().findElement(By.xpath("//a[@href= '/view/all/newJob']")).click();
+
+        getWait5().until(ExpectedConditions.visibilityOfElementLocated(By.id("name"))).sendKeys(jobName);
+        getDriver().findElement(By.className("org_jenkinsci_plugins_workflow_job_WorkflowJob")).click();
+        getWait2().until(ExpectedConditions.elementToBeClickable(By.id("ok-button"))).click();
+
+        getWait2().until(ExpectedConditions.elementToBeClickable(By.name("Submit"))).click();
+    }
+
+    private void goMainPageByBreadcrumb() {
+        getDriver().findElement(By.xpath("//div[@id = 'breadcrumbBar']//a[@href = '/']")).click();
+        getWait5().until(ExpectedConditions.numberOfElementsToBe(By.xpath("//li[@class='jenkins-breadcrumbs__list-item']"), 1));
+    }
+
+    private void runHelloWorldBuildInPipeline(String jobName) {
+        getDriver().findElement(By.xpath(CONFIGURE_ON_SIDE_PANEL_XPATH)).click();
+        getWait5().until(ExpectedConditions.textToBe(By.cssSelector("div#side-panel h1"), "Configure"));
+
+        Select select = new Select(getDriver().findElement(By.xpath("//div[@class='samples']/select")));
+        select.selectByValue("hello");
+        getWait2().until(ExpectedConditions.textToBePresentInElementLocated(By.xpath("//div[@class='ace_scroller']"), "Hello World"));
+
+        getDriver().findElement(By.name("Submit")).click();
+
+        getWait5().until(ExpectedConditions.elementToBeClickable(
+                By.xpath("//div[@id = 'tasks']//a[contains(@href, '/job/" + jobName + "/build')]"))).click();
+        getWait10().until(ExpectedConditions.visibilityOfElementLocated(By.xpath("//div[@class = 'table-box']")));
+    }
+    @Test(dependsOnMethods = "testDescriptionDisplays")
+    public void testDelete() {
+        getDriver().findElement((By.xpath(JOB_ON_DASHBOARD_XPATH))).click();
+        getWait2().until(ExpectedConditions.elementToBeClickable(By.xpath("//span[contains(text(),'Delete')]"))).click();
+
+        getWait2().until(ExpectedConditions.alertIsPresent()).accept();
+
+        getWait5().until(ExpectedConditions.numberOfElementsToBe(By.xpath("//li[@class='jenkins-breadcrumbs__list-item']"), 1));
+
+        Assert.assertEquals(getDriver().findElement(By.xpath("//h1")).getText(), "Welcome to Jenkins!");
+    }
+
+    @Test
+    public void testPipelineDeleteProject() {
+
+        getDriver().findElement(By.xpath("//a[@href = 'newJob']")).click();
+        getDriver().findElement(By.id("name")).sendKeys("newPipelineName");
+        getDriver().findElement(By.xpath("//span[contains(text(), 'Pipeline')]")).click();
+
+        getDriver().findElement(By.id("ok-button")).click();
+        getDriver().findElement(By.id("jenkins-name-icon")).click();
+
+        getDriver().findElement(By.xpath("//span[contains(text(), 'newPipelineName')]")).click();
+        getDriver().findElement(By.xpath("//span[contains(text(), 'Delete Pipeline')]")).click();
+
+        getDriver().switchTo().alert().accept();
+
+        String actualText = getDriver().findElement(By.xpath("//h1[contains(text(), 'Welcome to Jenkins!')]")).getText();
+
+        Assert.assertEquals(
+                actualText,
+                "Welcome to Jenkins!",
+                "Pipeline not deleted");
+    }
+
+    @Test
+    public void testCreate() {
+        createAPipeline(JOB_NAME);
+        goMainPageByBreadcrumb();
+
+        Assert.assertTrue(getDriver().findElement(By.xpath(JOB_ON_DASHBOARD_XPATH)).isDisplayed());
+        Assert.assertEquals(getDriver().findElement(By.xpath(JOB_ON_DASHBOARD_XPATH)).getText(), JOB_NAME);
+    }
+    @Test (dependsOnMethods = "testCreate")
+    public void testDescriptionDisplays() {
+        final String description = "Description of the Pipeline ";
+
+        getDriver().findElement(By.xpath(JOB_ON_DASHBOARD_XPATH)).click();
+        getDriver().findElement(By.id("description-link")).click();
+        getDriver().findElement(By.cssSelector("textarea[name ='description']")).sendKeys(description + JOB_NAME);
+        getDriver().findElement(By.xpath("//div[@id = 'description']//button[@name = 'Submit']")).click();
+
+        Assert.assertEquals(getWait2().until(ExpectedConditions.visibilityOfElementLocated(
+                        By.xpath("//div[@class = 'jenkins-buttons-row jenkins-buttons-row--invert']/preceding-sibling :: div"))).getText(),
+                description + JOB_NAME);
+    }
+
+    @Test
+    public void testPermalinksIsEmpty() {
+        final String jobName = "NewPipeline";
+
+        createAPipeline(jobName);
+        goMainPageByBreadcrumb();
+
+        getDriver().findElement(By.xpath("//td/a[@href='job/" + jobName + "/']")).click();
+
+        String permalinksInfo = getDriver().findElement(By.xpath("//ul[@class = 'permalinks-list']")).getText();
+
+        Assert.assertTrue(permalinksInfo.isEmpty());
+    }
+
+
+    @Test
+    public void testPermalinksContainBuildInformation() throws InterruptedException {
+        final String jobName = "Pipeline2";
+        final List<String> buildsInfo = List.of("Last build (#1)", "Last stable build (#1)", "Last successful build (#1)",
+                "Last completed build (#1)");
+
+        createAPipeline(jobName);
+        goMainPageByBreadcrumb();
+
+        getDriver().findElement(By.xpath("//td//a[@title = 'Schedule a Build for " + jobName + "']")).click();
+        Thread.sleep(2000);
+
+        getDriver().findElement(By.xpath("//td/a[@href='job/" + jobName + "/']")).click();
+
+        List<WebElement> permalinks = getDriver().findElements(By.cssSelector(".permalink-item"));
+
+        Assert.assertEquals(permalinks.size(), 4);
+        for (int i = 0; i < permalinks.size(); i++) {
+            Assert.assertTrue(permalinks.get(i).getText().contains(buildsInfo.get(i)));
+        }
+    }
+
+    @Test
+    public void testStageViewBeforeBuild() {
+        createAPipeline(JOB_NAME);
+        goMainPageByBreadcrumb();
+
+        getDriver().findElement(By.xpath(JOB_ON_DASHBOARD_XPATH)).click();
+
+        String stageViewInfo = getDriver().findElement(By.cssSelector("div#pipeline-box > div")).getText();
+
+        Assert.assertEquals(stageViewInfo, "No data available. This Pipeline has not yet run.");
+    }
+
+    @Test
+    public void testPermalinksContainsInfo() throws InterruptedException {
+        final String pipelineName = "Pipeline_Test";
+        final List<String> permalinksInfo = List.of(
+                "Last build (#1)",
+                "Last stable build (#1)",
+                "Last successful build (#1)",
+                "Last completed build (#1)"
+        );
+
+        createPipeline(pipelineName);
+
+        getDriver().findElement(By.xpath("//td//a[@title = 'Schedule a Build for " + pipelineName + "']")).click();
+        Thread.sleep(2000);
+
+        getDriver().findElement(By.xpath("//td//a[@href = 'job/" + pipelineName + "/']")).click();
+
+        List<WebElement> permalinks = getDriver().findElements(By.className("permalink-item"));
+
+        Assert.assertEquals(permalinks.size(), 4);
+        for (int i = 0; i < permalinks.size(); i++) {
+            Assert.assertTrue(permalinks.get(i).getText().contains(permalinksInfo.get(i)));
+        }
+    }
+
+    @Test (dependsOnMethods = "testStageViewBeforeBuild")
+    public void testStageViewAfterRunningSampleBuild() {
+        getDriver().findElement(By.xpath(JOB_ON_DASHBOARD_XPATH)).click();
+        runHelloWorldBuildInPipeline(JOB_NAME);
+
+        Assert.assertTrue(getWait2().until(ExpectedConditions.visibilityOfElementLocated(
+                By.xpath("//div[@class = 'table-box']"))).isDisplayed());
+        Assert.assertEquals(getDriver().findElement(
+                        By.xpath("//table[@class = 'jobsTable']//th[@class = 'stage-header-name-0']")).getText(),
+                "Hello");
+    }
+
+    @Test
+    public void testSaveSettingsWhileConfigure() {
+        createAPipeline(JOB_NAME);
+        goMainPageByBreadcrumb();
+
+        getDriver().findElement(By.xpath(JOB_ON_DASHBOARD_XPATH)).click();
+        getDriver().findElement(By.xpath(CONFIGURE_ON_SIDE_PANEL_XPATH)).click();
+        getWait5().until(ExpectedConditions.elementToBeClickable(
+                By.xpath("//label[contains(text(), '" + CHECKBOX_TEXT + "')]"))).click();
+        getDriver().findElement(By.name("Submit")).click();
+
+        getDriver().findElement(By.xpath(CONFIGURE_ON_SIDE_PANEL_XPATH)).click();
+
+        Assert.assertTrue(getWait5().until(ExpectedConditions.presenceOfElementLocated(
+                By.xpath("//label[contains(text(), '" + CHECKBOX_TEXT + "')]/../input"))).isSelected());
+    }
+
+    @Test(dependsOnMethods = "testSaveSettingsWhileConfigure")
+    public void testUnsavedSettingsWhileConfigure() {
+        getDriver().findElement(By.xpath(JOB_ON_DASHBOARD_XPATH)).click();
+        getDriver().findElement(By.xpath(CONFIGURE_ON_SIDE_PANEL_XPATH)).click();
+        getWait5().until(ExpectedConditions.elementToBeClickable(
+                By.xpath("//label[contains(text(), '" + CHECKBOX_TEXT + "')]"))).click();
+        getDriver().findElement(By.id("jenkins-name-icon")).click();
+
+        getWait2().until(ExpectedConditions.alertIsPresent()).accept();
+
+        getDriver().findElement(By.xpath(JOB_ON_DASHBOARD_XPATH)).click();
+        getDriver().findElement(By.xpath(CONFIGURE_ON_SIDE_PANEL_XPATH)).click();
+
+        Assert.assertTrue(getWait5().until(ExpectedConditions.presenceOfElementLocated(
+                By.xpath("//label[contains(text(), '" + CHECKBOX_TEXT + "')]/../input"))).isSelected());
+    }
+
+    @Test
+    public void testTooltipsDescriptionCompliance() {
+        final String jobName = "Another_Pipeline";
+
+        createAPipeline(jobName);
+        goMainPageByBreadcrumb();
+
+        getDriver().findElement(By.xpath("//tr[@id ='job_" + jobName + "']//a[@href = 'job/" + jobName + "/']")).click();
+        getDriver().findElement(By.xpath("//div[@id = 'tasks']//a[@href = '/job/" + jobName + "/configure']")).click();
+
+        List<WebElement> toolTips = getDriver().findElements(By.xpath("//div[@hashelp = 'true']//a[contains(@tooltip, '')]"));
+        List<WebElement> checkBoxesWithTooltips = getDriver().findElements(By.xpath("//div[@hashelp = 'true']//label[@class = 'attach-previous ']"));
+
+        Assert.assertEquals(toolTips.size(), 11);
+        Assert.assertEquals(toolTips.size(), checkBoxesWithTooltips.size());
+        for (int i = 0; i < toolTips.size(); i++) {
+            Assert.assertEquals("Help for feature: " + checkBoxesWithTooltips.get(i).getText(), toolTips.get(i).getAttribute("title"));
+        }
+    }
+    @Test
+    public void testPermalinksBuildData() {
+        final String jobName = "Pipeline1";
+
+        createAPipeline(jobName);
+        goMainPageByBreadcrumb();
+
+        getDriver().findElement(By.xpath("//td/a[@href='job/" + jobName + "/']")).click();
+        getDriver().findElement(By.xpath("//a[@href='/job/" + jobName + "/build?delay=0sec']")).click();
+        getWait5().until(ExpectedConditions.visibilityOfElementLocated(By.xpath("//div[@class='alert alert-warning']")));
+
+        getDriver().navigate().refresh();
+        List<WebElement> permalinksBuildHistory = getWait5().until(ExpectedConditions.presenceOfAllElementsLocatedBy(By.xpath("//li[@class='permalink-item']")));
+
+        Assert.assertEquals(permalinksBuildHistory.size(), 4);
+        Assert.assertTrue(permalinksBuildHistory.get(0).getText().contains("Last build"));
+        Assert.assertTrue(permalinksBuildHistory.get(1).getText().contains("Last stable build"));
+        Assert.assertTrue(permalinksBuildHistory.get(2).getText().contains("Last successful build"));
+        Assert.assertTrue(permalinksBuildHistory.get(3).getText().contains("Last completed build"));
     }
 }
